moved surface code from wayland_connector.cpp to wl_surface.cpp
