+++ src/server/frontend/wayland/wayland_connector.cpp
@@ -18,6 +18,8 @@
 
 #include "wayland_connector.h"
 
+#include "wl_surface.h"
+
 #include "basic_surface_event_sink.h"
 #include "null_event_sink.h"
 #include "output_manager.h"
@@ -87,6 +89,18 @@ namespace mir
 {
 namespace frontend
 {
+struct NullWlMirWindow : WlMirWindow
+{
+    void new_buffer_size(mir::geometry::Size const& ) {}
+    void commit() {}
+    void visiblity(bool ) {}
+    void destroy() {}
+
+    static NullWlMirWindow instance;
+} NullWlMirWindow::instance;
+
+WlMirWindow * const nullWlMirWindowPtr = &NullWlMirWindow::instance;
+
 namespace
 {
 struct ClientPrivate
@@ -130,16 +144,6 @@ ClientPrivate* private_from_listener(wl_listener* listener)
 void cleanup_private(wl_listener* listener, void* /*data*/)
 {
     delete private_from_listener(listener);
-}
-
-std::shared_ptr<mf::Session> session_for_client(wl_client* client)
-{
-    auto listener = wl_client_get_destroy_listener(client, &cleanup_private);
-
-    if (listener)
-        return private_from_listener(listener)->session;
-
-    return nullptr;
 }
 
 std::shared_ptr<ms::Surface> get_surface_for_id(std::shared_ptr<mf::Session> const& session, mf::SurfaceId surface_id)
@@ -231,252 +235,16 @@ std::shared_ptr<mf::BufferStream> create_buffer_stream(mf::Session& session)
     return session.get_buffer_stream(id);
 }
 */
-template<typename Callable>
-auto run_unless(std::shared_ptr<bool> const& condition, Callable&& callable)
-{
-    return
-        [callable = std::move(callable), condition]()
-        {
-            if (*condition)
-                return;
-
-            callable();
-        };
-}
 }
 
-struct WlMirWindow
+std::shared_ptr<mir::frontend::Session> session_for_client(wl_client* client)
 {
-    virtual void new_buffer_size(geometry::Size const& buffer_size) = 0;
-    virtual void commit() = 0;
-    virtual void visiblity(bool visible) = 0;
-    virtual void destroy() = 0;
-    virtual ~WlMirWindow() = default;
-};
+    auto listener = wl_client_get_destroy_listener(client, &cleanup_private);
 
-struct NullWlMirWindow : WlMirWindow
-{
-    void new_buffer_size(geom::Size const& ) {}
-    void commit() {}
-    void visiblity(bool ) {}
-    void destroy() {}
-} nullwlmirwindow;
+    if (listener)
+        return private_from_listener(listener)->session;
 
-class WlSurface : public wayland::Surface
-{
-public:
-    WlSurface(
-        wl_client* client,
-        wl_resource* parent,
-        uint32_t id,
-        std::shared_ptr<mir::Executor> const& executor,
-        std::shared_ptr<mg::WaylandAllocator> const& allocator)
-        : Surface(client, parent, id),
-          allocator{allocator},
-          executor{executor},
-          pending_buffer{nullptr},
-          pending_frames{std::make_shared<std::vector<wl_resource*>>()},
-          destroyed{std::make_shared<bool>(false)}
-    {
-        auto session = session_for_client(client);
-        mg::BufferProperties const props{
-            geom::Size{geom::Width{0}, geom::Height{0}},
-            mir_pixel_format_invalid,
-            mg::BufferUsage::undefined
-        };
-
-        stream_id = session->create_buffer_stream(props);
-        stream = session->get_buffer_stream(stream_id);
-
-        // wl_surface is specified to act in mailbox mode
-        stream->allow_framedropping(true);
-    }
-
-    ~WlSurface()
-    {
-        *destroyed = true;
-        if (auto session = session_for_client(client))
-            session->destroy_buffer_stream(stream_id);
-    }
-
-    void set_role(WlMirWindow* role_)
-    {
-        role = role_;
-    }
-
-    mf::BufferStreamId stream_id;
-    std::shared_ptr<mf::BufferStream> stream;
-    mf::SurfaceId surface_id;       // ID of any associated surface
-
-    std::shared_ptr<bool> destroyed_flag() const
-    {
-        return destroyed;
-    }
-
-private:
-    std::shared_ptr<mg::WaylandAllocator> const allocator;
-    std::shared_ptr<mir::Executor> const executor;
-
-    WlMirWindow* role = &nullwlmirwindow;
-
-    wl_resource* pending_buffer;
-    std::shared_ptr<std::vector<wl_resource*>> const pending_frames;
-    std::shared_ptr<bool> const destroyed;
-
-    void destroy();
-    void attach(std::experimental::optional<wl_resource*> const& buffer, int32_t x, int32_t y);
-    void damage(int32_t x, int32_t y, int32_t width, int32_t height);
-    void frame(uint32_t callback);
-    void set_opaque_region(std::experimental::optional<wl_resource*> const& region);
-    void set_input_region(std::experimental::optional<wl_resource*> const& region);
-    void commit();
-    void damage_buffer(int32_t x, int32_t y, int32_t width, int32_t height);
-    void set_buffer_transform(int32_t transform);
-    void set_buffer_scale(int32_t scale);
-};
-
-WlSurface* get_wlsurface(wl_resource* surface)
-{
-    auto* tmp = wl_resource_get_user_data(surface);
-    return static_cast<WlSurface*>(static_cast<wayland::Surface*>(tmp));
-}
-
-void WlSurface::destroy()
-{
-    *destroyed = true;
-    role->destroy();
-    wl_resource_destroy(resource);
-}
-
-void WlSurface::attach(std::experimental::optional<wl_resource*> const& buffer, int32_t x, int32_t y)
-{
-    if (x != 0 || y != 0)
-    {
-        mir::log_warning("Client requested unimplemented non-zero attach offset. Rendering will be incorrect.");
-    }
-
-    role->visiblity(!!buffer);
-
-    pending_buffer = buffer.value_or(nullptr);
-}
-
-void WlSurface::damage(int32_t x, int32_t y, int32_t width, int32_t height)
-{
-    (void)x;
-    (void)y;
-    (void)width;
-    (void)height;
-    // This isn't essential, but could enable optimizations
-}
-
-void WlSurface::damage_buffer(int32_t x, int32_t y, int32_t width, int32_t height)
-{
-    (void)x;
-    (void)y;
-    (void)width;
-    (void)height;
-    // This isn't essential, but could enable optimizations
-}
-
-void WlSurface::frame(uint32_t callback)
-{
-    pending_frames->emplace_back(
-        wl_resource_create(client, &wl_callback_interface, 1, callback));
-}
-
-void WlSurface::set_opaque_region(const std::experimental::optional<wl_resource*>& region)
-{
-    (void)region;
-}
-
-void WlSurface::set_input_region(const std::experimental::optional<wl_resource*>& region)
-{
-    (void)region;
-}
-
-void WlSurface::commit()
-{
-    if (pending_buffer)
-    {
-        auto send_frame_notifications =
-            [executor = executor, frames = pending_frames, destroyed = destroyed]()
-            {
-                executor->spawn(run_unless(
-                    destroyed,
-                    [frames]()
-                    {
-                        /*
-                         * There is no synchronisation required here -
-                         * This is run on the WaylandExecutor, and is guaranteed to run on the
-                         * wl_event_loop's thread.
-                         *
-                         * The only other accessors of WlSurface are also on the wl_event_loop,
-                         * so this is guaranteed not to be reentrant.
-                         */
-                        for (auto frame : *frames)
-                        {
-                            wl_callback_send_done(frame, 0);
-                            wl_resource_destroy(frame);
-                        }
-                        frames->clear();
-                    }));
-            };
-
-        std::shared_ptr<mg::Buffer> mir_buffer;
-
-        if (wl_shm_buffer_get(pending_buffer))
-        {
-            mir_buffer = WlShmBuffer::mir_buffer_from_wl_buffer(
-                pending_buffer,
-                std::move(send_frame_notifications));
-        }
-        else
-        {
-            auto release_buffer = [executor = executor, buffer = pending_buffer, destroyed = destroyed]()
-                {
-                    executor->spawn(run_unless(
-                        destroyed,
-                        [buffer](){ wl_resource_queue_event(buffer, WL_BUFFER_RELEASE); }));
-                };
-
-            mir_buffer = allocator->buffer_from_resource(
-                    pending_buffer,
-                    std::move(send_frame_notifications),
-                    std::move(release_buffer));
-        }
-
-        /*
-         * This is technically incorrect - the resize and submit_buffer *should* be atomic,
-         * but are not, so a client in the process of resizing can have buffers rendered at
-         * an unexpected size.
-         *
-         * It should be good enough for now, though.
-         *
-         * TODO: Provide a mg::Buffer::logical_size() to do this properly.
-         */
-        stream->resize(mir_buffer->size());
-        role->new_buffer_size(mir_buffer->size());
-        role->commit();
-        stream->submit_buffer(mir_buffer);
-
-        pending_buffer = nullptr;
-    }
-    else
-    {
-        role->commit();
-    }
-}
-
-void WlSurface::set_buffer_transform(int32_t transform)
-{
-    (void)transform;
-    // TODO
-}
-
-void WlSurface::set_buffer_scale(int32_t scale)
-{
-    (void)scale;
-    // TODO
+    return nullptr;
 }
 
 class WlCompositor : public wayland::Compositor
@@ -624,7 +392,7 @@ public:
                 destroyed,
                 [
                     target = target,
-                    target_window_destroyed = get_wlsurface(target)->destroyed_flag(),
+                    target_window_destroyed = WlSurface::from(target)->destroyed_flag(),
                     focussed = mir_window_event_get_attribute_value(event),
                     this
                 ]()
@@ -827,7 +595,7 @@ public:
             [
                 ev = mcl::Event{mir_input_event_get_event(event)},
                 target,
-                target_window_destroyed = get_wlsurface(target)->destroyed_flag(),
+                target_window_destroyed = WlSurface::from(target)->destroyed_flag(),
                 this
             ]()
             {
@@ -990,7 +758,7 @@ public:
             [
                 ev = mcl::Event{mir_input_event_get_event(event)},
                 target = target,
-                target_window_destroyed = get_wlsurface(target)->destroyed_flag(),
+                target_window_destroyed = WlSurface::from(target)->destroyed_flag(),
                 this
             ]()
             {
@@ -1511,7 +1279,7 @@ private:
         if (latest_buffer_size == geom::Size{})
             return;
 
-        auto* const mir_surface = get_wlsurface(surface);
+        auto* const mir_surface = WlSurface::from(surface);
         if (params.size == geom::Size{}) params.size = latest_buffer_size;
         params.content_id = mir_surface->stream_id;
         surface_id = shell->create_surface(session, params, sink);
@@ -1568,8 +1336,8 @@ public:
 
     ~WlShellSurface() override
     {
-        auto* const mir_surface = get_wlsurface(surface);
-        mir_surface->set_role(&nullwlmirwindow);
+        auto* const mir_surface = WlSurface::from(surface);
+        mir_surface->set_role(nullWlMirWindowPtr);
     }
 
 protected:
@@ -1580,7 +1348,7 @@ protected:
 
     void set_toplevel() override
     {
-        auto* const mir_surface = get_wlsurface(surface);
+        auto* const mir_surface = WlSurface::from(surface);
 
         mir_surface->set_role(this);
     }
@@ -1592,7 +1360,7 @@ protected:
         uint32_t flags) override
     {
         auto const session = session_for_client(client);
-        auto& parent_surface = *get_wlsurface(parent);
+        auto& parent_surface = *WlSurface::from(parent);
 
         if (surface_id.as_value())
         {
@@ -1617,7 +1385,7 @@ protected:
             params.aux_rect_placement_offset_x = 0;
             params.aux_rect_placement_offset_y = 0;
 
-            auto* const mir_surface = get_wlsurface(surface);
+            auto* const mir_surface = WlSurface::from(surface);
             mir_surface->set_role(this);
         }
     }
@@ -1655,7 +1423,7 @@ protected:
         uint32_t flags) override
     {
         auto const session = session_for_client(client);
-        auto& parent_surface = *get_wlsurface(parent);
+        auto& parent_surface = *WlSurface::from(parent);
 
         if (surface_id.as_value())
         {
@@ -1681,7 +1449,7 @@ protected:
             params.aux_rect_placement_offset_x = 0;
             params.aux_rect_placement_offset_y = 0;
 
-            auto* const mir_surface = get_wlsurface(surface);
+            auto* const mir_surface = WlSurface::from(surface);
             mir_surface->set_role(this);
         }
     }
@@ -2004,8 +1772,8 @@ struct XdgSurfaceV6 : wayland::XdgSurfaceV6, WlAbstractMirWindow
 
     ~XdgSurfaceV6() override
     {
-        auto* const mir_surface = get_wlsurface(surface);
-        mir_surface->set_role(&nullwlmirwindow);
+        auto* const mir_surface = WlSurface::from(surface);
+        mir_surface->set_role(nullWlMirWindowPtr);
     }
 
     void destroy() override
@@ -2016,7 +1784,7 @@ struct XdgSurfaceV6 : wayland::XdgSurfaceV6, WlAbstractMirWindow
     void get_toplevel(uint32_t id) override
     {
         new XdgToplevelV6{client, parent, id, shell, this};
-        auto* const mir_surface = get_wlsurface(surface);
+        auto* const mir_surface = WlSurface::from(surface);
         mir_surface->set_role(this);
     }
 
@@ -2039,7 +1807,7 @@ struct XdgSurfaceV6 : wayland::XdgSurfaceV6, WlAbstractMirWindow
         params.placement_hints = mir_placement_hints_slide_any;
 
         new XdgPopupV6{client, parent, id};
-        auto* const mir_surface = get_wlsurface(surface);
+        auto* const mir_surface = WlSurface::from(surface);
         mir_surface->set_role(this);
     }
