+++ src/server/frontend/wayland/wayland_connector.h
@@ -28,7 +28,6 @@
 
 namespace mir
 {
-
 namespace input
 {
 class InputDeviceHub;
@@ -39,7 +38,10 @@ namespace graphics
 class GraphicBufferAllocator;
 class WaylandAllocator;
 }
-
+namespace geometry
+{
+struct Size;
+}
 namespace frontend
 {
 class WlCompositor;
@@ -53,6 +55,30 @@ class Shell;
 class DisplayChanger;
 class SessionAuthorizer;
 class DataDeviceManager;
+
+template<typename Callable>
+inline auto run_unless(std::shared_ptr<bool> const& condition, Callable&& callable)
+{
+    return
+        [callable = std::move(callable), condition]()
+        {
+            if (*condition)
+                return;
+
+            callable();
+        };
+}
+
+std::shared_ptr<mir::frontend::Session> session_for_client(wl_client* client);
+
+struct WlMirWindow
+{
+    virtual void new_buffer_size(geometry::Size const& buffer_size) = 0;
+    virtual void commit() = 0;
+    virtual void visiblity(bool visible) = 0;
+    virtual void destroy() = 0;
+    virtual ~WlMirWindow() = default;
+} extern * const nullWlMirWindowPtr;
 
 class WaylandConnector : public Connector
 {
@@ -78,6 +104,7 @@ public:
         std::function<void(std::shared_ptr<Session> const& session)> const& connect_handler) const override;
 
     void run_on_wayland_display(std::function<void(wl_display*)> const& functor);
+
 private:
     std::unique_ptr<wl_display, void(*)(wl_display*)> const display;
     mir::Fd const pause_signal;
@@ -91,8 +118,6 @@ private:
     std::thread dispatch_thread;
     wl_event_source* pause_source;
 };
-
-
 }
 }
